Use systemtime for postgres datetime type
